WIZ-28: add types to handleServerError, etc.
