+++ packages/client/package.json
@@ -65,7 +65,7 @@
     "stylelint-config-standard-scss": "7.0.1",
     "stylelint-order": "6.0.3",
     "ts-jest": "28.0.8",
-    "typescript": "4.8.2",
+    "typescript": "5.0.3",
     "vite": "3.0.7"
   },
   "license": "MIT"
