+++ packages/client/src/features/Profile/ProfileChangeAvatar/ProfileChangeAvatar.scss
@@ -1,5 +1,3 @@
-@import 'core/vars';
-
 .upload-container {
   margin: 8px 0 0;
   text-align: center;
