+++ packages/client/src/features/Profile/ProfileChangeAvatar/common.ts
@@ -1,4 +1,6 @@
-import { defineMessages } from 'react-intl';
+import { type IntlFormatters, defineMessages } from 'react-intl';
+import { message } from 'antd';
+import type { RcFile } from 'antd/es/upload';
 
 export const messages = defineMessages({
   textUpload: { id: 'universal.upload', defaultMessage: 'Upload' },
@@ -11,3 +13,29 @@ export const messages = defineMessages({
     defaultMessage: 'Image must smaller than 2MB',
   },
 });
+
+type BeforeUploadOptions = {
+  fm: IntlFormatters['formatMessage'];
+};
+
+export async function beforeUpload(file: RcFile, { fm }: BeforeUploadOptions) {
+  const isJpgOrPng = file.type === 'image/jpeg' || file.type === 'image/png';
+
+  if (!isJpgOrPng) {
+    await message.open({
+      type: 'error',
+      content: fm(messages.validationAvatarInvalidExtension),
+    });
+  }
+
+  const isLt2M = file.size / 1024 / 1024 < 2;
+
+  if (!isLt2M) {
+    await message.open({
+      type: 'error',
+      content: fm(messages.validationAvatarSizeLimit),
+    });
+  }
+
+  return isJpgOrPng && isLt2M;
+}
