+++ packages/client/src/features/Profile/ProfileChangePassword/ProfileChangePassword.scss
@@ -1,5 +1,3 @@
-@import 'core/vars';
-
 .form-change-password {
   display: flex;
   flex-direction: column;
