+++ packages/client/src/features/Profile/ProfileChangePassword/ProfileChangePassword.tsx
@@ -44,7 +44,7 @@ export const ProfileChangePassword: FC = () => {
         navigate(ROUTES.PROFILE_PAGE.path);
       }
     } catch (err) {
-      await handleServerError(err);
+      await handleServerError(err as ServerError);
     }
 
     setIsSubmitting(false);
