+++ packages/client/src/features/Profile/ProfileForm/ProfileForm.scss
@@ -1,5 +1,3 @@
-@import 'core/vars';
-
 .form-profile {
   display: flex;
   flex-direction: column;
