+++ packages/client/src/helpers/handleServerError.ts
@@ -1,12 +1,10 @@
 import { message } from 'antd';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export async function handleServerError(err: any) {
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, no-console
+export async function handleServerError(err: ServerError) {
   console.error({ err });
+
   await message.open({
     type: 'error',
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
-    content: err.response?.data?.reason,
+    content: err.response?.data.reason,
   });
 }
