+++ packages/client/src/types/app.d.ts
@@ -1,6 +1,9 @@
-export {};
+import { AxiosError } from 'axios';
 
 declare global {
   type LocaleType = 'en' | 'ru';
   type Nullable<T> = T | null;
+  type ServerError = AxiosError<{ reason: string }>;
 }
+
+export {};
